requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
